Move context to src/
