+++ src/context/Providers.tsx
@@ -3,8 +3,8 @@ import { BrowserRouter as Router } from "react-router-dom";
 import { ThemeProvider } from "theme-ui";
 import { Provider as UrqlProvider } from "urql";
 
-import { client } from "../../graphql/client";
-import { theme } from "../../theme";
+import { client } from "../graphql/client";
+import { theme } from "../theme";
 
 export const Providers: FC = ({ children }) => (
   <React.StrictMode>
