+++ src/index.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from "react-dom";
 
 import { App } from "./App";
-import { Providers } from "./components/context/Providers";
+import { Providers } from "./context/Providers";
 
 ReactDOM.render(
   <Providers>
